+++ core/tabs/applications-setup/design-tools/handbrake.sh
@@ -4,7 +4,10 @@
 
 installHandbrake() {
 	printf "%b\n" "${YELLOW}Installing Handbrake...${RC}"
-	if ! command_exists handbrake; then
+	if ! flatpak_app_installed fr.handbrake.ghb && ! command_exists handbrake; then
+	    if try_flatpak_install fr.handbrake.ghb; then
+	        return 0
+	    fi
 	    case "$PACKAGER" in
 	        apt-get|nala)
 				"$ESCALATION_TOOL" "$PACKAGER" install -y handbrake
@@ -13,10 +16,8 @@ installHandbrake() {
 			    "$AUR_HELPER" -S --needed --noconfirm --cleanafter handbrake
 	            ;;
 	        *)
-	        	if command_exists flatpak; then
-	            	"$ESCALATION_TOOL" flatpak install --noninteractive fr.handbrake.ghb
-	            fi
-	            exit 1
+	        	printf "%b\n" "${RED}Flatpak install failed and no native package is configured for ${PACKAGER}.${RC}"
+	        	exit 1
 	            ;;
 	    esac
 	else
@@ -26,6 +27,9 @@ installHandbrake() {
 
 uninstallHandbrake() {
 	printf "%b\n" "${YELLOW}Uninstalling Handbrake...${RC}"
+	if uninstall_flatpak_if_installed fr.handbrake.ghb; then
+	    return 0
+	fi
 	if command_exists handbrake; then
 	    case "$PACKAGER" in
 	        apt-get|nala|dnf|zypper)
@@ -35,7 +39,7 @@ uninstallHandbrake() {
 			    "$AUR_HELPER" -R --noconfirm --cleanafter handbrake
 	            ;;
 	        *)
-	            "$ESCALATION_TOOL" flatpak uninstall --noninteractive fr.handbrake.ghb
+	            printf "%b\n" "${RED}No native uninstall is configured for ${PACKAGER}.${RC}"
 	            exit 1
 	            ;;
 	    esac
