+++ core/tabs/applications-setup/design-tools/mypaint.sh
@@ -4,7 +4,10 @@
 
 installMyPaint() {
 	printf "%b\n" "${YELLOW}Installing MyPaint...${RC}"
-	if ! command_exists mypaint; then
+	if ! flatpak_app_installed org.mypaint.MyPaint && ! command_exists mypaint; then
+	    if try_flatpak_install org.mypaint.MyPaint; then
+	        return 0
+	    fi
 	    case "$PACKAGER" in
 	        apt-get|nala|dnf|zypper)
 				"$ESCALATION_TOOL" "$PACKAGER" install -y mypaint
@@ -13,10 +16,8 @@ installMyPaint() {
 			    "$AUR_HELPER" -S --needed --noconfirm --cleanafter mypaint
 	            ;;
 	        *)
-	        	if command_exists flatpak; then
-	            	"$ESCALATION_TOOL" flatpak install --noninteractive org.mypaint.MyPaint
-	            fi
-	            exit 1
+	        	printf "%b\n" "${RED}Flatpak install failed and no native package is configured for ${PACKAGER}.${RC}"
+	        	exit 1
 	            ;;
 	    esac
 	else
@@ -26,6 +27,9 @@ installMyPaint() {
 
 uninstallMyPaint() {
 	printf "%b\n" "${YELLOW}Uninstalling MyPaint...${RC}"
+	if uninstall_flatpak_if_installed org.mypaint.MyPaint; then
+	    return 0
+	fi
 	if command_exists mypaint; then
 	    case "$PACKAGER" in
 	        apt-get|nala|dnf|zypper)
@@ -35,7 +39,7 @@ uninstallMyPaint() {
 			    "$AUR_HELPER" -R --noconfirm --cleanafter mypaint
 	            ;;
 	        *)
-	            "$ESCALATION_TOOL" flatpak uninstall --noninteractive org.mypaint.mypaint
+	            printf "%b\n" "${RED}No native uninstall is configured for ${PACKAGER}.${RC}"
 	            exit 1
 	            ;;
 	    esac
